requirements.txt
asgiref==3.7.2
Django==4.2.2
django-environ==0.10.0
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-nested-routers==0.93.4
flake8==6.1.0
mccabe==0.7.0
psycopg2==2.9.6
pycodestyle==2.11.0
pyflakes==3.1.0
PyJWT==2.8.0
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
